feat: Add mongo-id pipe and update related files
The commit adds a new `mongo-id.pipe.ts` file to the `common/pipes` directory. This pipe is responsible for validating if a given value is a valid MongoDB ID. It throws a `BadRequestException` if the value is not a valid ID.

The following files were updated to include the new pipe:
- `common/index.ts` to export the `mongo-id.pipe.ts`
- `auth/application/dtos/index.ts` to export the `update-user.dto.ts`
- `auth/application/use-cases/index.ts` to export the `update-user.use-case.ts`
- `auth/application/interfaces/auth-service.interface.ts` to include the `updateUser` method
- `auth/auth.module.ts` to include the `UpdateUserUseCase` in the module's providers
- `users/application/interfaces/user-repository.interface.ts` to include the `update` method
- `users/application/dtos/update-user.dto.ts` to define the `UpdateUserDto` class
- `users/application/use-cases/update-user.use-case.ts` to implement the `UpdateUserUseCase`
- `users/infrastructure/repositories/user.repository.ts` to implement the `update` method

This commit adds the necessary functionality to validate and update user information using the MongoDB ID pipe and related files.