feat: Add product code to Product entity and related files
This commit adds the `productCode` property to the `Product` entity in the `product.entity.ts` file. It also updates the `ProductMapper`, `CreateProductDto`, and `ProductResponseDto` files to include the `productCode` property. Additionally, the `ProductRepository` is modified to validate the uniqueness of the `productCode` when creating a new product. This enhancement improves the functionality of the application by allowing products to be identified by their unique product codes.